polynew-main/requirements.txt
httpx>=0.27.0
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.0
flask>=3.0.0